Extract non-async tests
