+++ test/sync_/test_match_api.py
@@ -31,9 +31,7 @@ from neomodel.sync_.match import (
     RawCypher,
     RelationNameResolver,
     Size,
-    Traversal,
 )
-from neomodel.util import RelationshipDirection
 
 
 class SupplierRel(StructuredRel):
@@ -413,33 +411,6 @@ def test_extra_filters():
         Coffee.nodes.filter(elementId="4:xxx:111").all()
 
 
-def test_traversal_definition_keys_are_valid():
-    muckefuck = Coffee(name="Mukkefuck", price=1)
-
-    with raises(ValueError):
-        Traversal(
-            muckefuck,
-            "a_name",
-            {
-                "node_class": Supplier,
-                "direction": RelationshipDirection.INCOMING,
-                "relationship_type": "KNOWS",
-                "model": None,
-            },
-        )
-
-    Traversal(
-        muckefuck,
-        "a_name",
-        {
-            "node_class": Supplier,
-            "direction": RelationshipDirection.INCOMING,
-            "relation_type": "KNOWS",
-            "model": None,
-        },
-    )
-
-
 @mark_sync_test
 def test_empty_filters():
     """Test this case:
@@ -571,18 +542,6 @@ def test_q_filters():
 
     with raises(TypeError):
         Coffee.nodes.filter(Q(price=5) | QQ()).all()
-
-
-def test_qbase():
-    test_print_out = str(Q(price=5) | Q(price=10))
-    test_repr = repr(Q(price=5) | Q(price=10))
-    assert test_print_out == "(OR: ('price', 5), ('price', 10))"
-    assert test_repr == "<Q: (OR: ('price', 5), ('price', 10))>"
-
-    assert ("price", 5) in (Q(price=5) | Q(price=10))
-
-    test_hash = set([Q(price_lt=30) | ~Q(price=5), Q(price_lt=30) | ~Q(price=5)])
-    assert len(test_hash) == 1
 
 
 @mark_sync_test
